fix(ergo-ser): AvlTreeData keyLength/valueLengthOpt wrap to signed i32 (Scala parity)
Scala AvlTreeData.parse reads `keyLength = r.getUInt().toInt` and
`valueLengthOpt = r.getOption(r.getUInt().toInt)` with case-class fields
`keyLength: Int` / `valueLengthOpt: Option[Int]`: a length encoding above
i32::MAX WRAPS to a negative Int and the deserializer SUCCEEDS with
invalid-but-parseable AvlTreeData (its own comment: "when keyLength and
valueLengthOpt < 0 as a result of Int overflow, the deserializer succeeds").
Our reader used get_u32_exact (bounded to i32::MAX), so such a tree was
REJECTED at parse, erroring the whole tree.

- ergo-primitives: add VlqReader::get_uint_to_i32 (Scala getUInt().toInt:
  VLQ-decode bounded to u32::MAX, then wrap to i32).
- ergo-ser: AvlTreeData.key_length/value_length_opt u32 -> i32; read via
  get_uint_to_i32; write via `x as u32` (round-trips wrapped bytes).
- ergo-sigma: keyLength/valueLengthOpt accessors return the i32 directly
  (a wrapped tree's keyLength surfaces as a negative SInt); make_avl_verifier
  rejects invalid metadata (keyLength<=0 || valueLengthOpt<0) up front,
  mirroring scrypto require(keyLength>0) and avoiding a huge `as usize`
  (try_make_avl_verifier maps it to None -> contains->false, remove->None,
  etc.); avl_cost_height uses the signed view directly.
- ergo-validation: ERGO_UTXO_KEY_LENGTH u32 -> i32.

Closes avl-keylength-uint-exact-rejected (keyLength/valueLengthOpt-negative
accessors + the negative-keyLength trees' contains/remove/insert + the
AvlTree_properties contains entries).

Test: read_avl_tree_wraps_lengths_above_i32_max (keyLength 2^31 -> i32::MIN,
valueLengthOpt 2^32-1 -> -1).

+++ ergo-sigma/src/evaluator/opcodes/property_call.rs
@@ -231,7 +231,10 @@ pub(super) fn eval_no_arg_method(
         (100, 3) => {
             add_method_cost(cost, 15)?;
             match obj_val {
-                Value::AvlTree(avl) => Ok(Some(Value::Int(avl.key_length as i32))),
+                // key_length is i32 (Scala keyLength: Int): a wrapped-negative
+                // length surfaces as a negative SInt here, matching the
+                // reference (e.g. keyLength-negative -> Int(-2147483648)).
+                Value::AvlTree(avl) => Ok(Some(Value::Int(avl.key_length))),
                 other => Err(EvalError::TypeError {
                     expected: "AvlTree for keyLength",
                     got: format!("{other:?}"),
@@ -242,7 +245,7 @@ pub(super) fn eval_no_arg_method(
             add_method_cost(cost, 15)?;
             match obj_val {
                 Value::AvlTree(avl) => Ok(Some(Value::Opt(
-                    avl.value_length_opt.map(|v| Box::new(Value::Int(v as i32))),
+                    avl.value_length_opt.map(|v| Box::new(Value::Int(v))),
                 ))),
                 other => Err(EvalError::TypeError {
                     expected: "AvlTree for valueLengthOpt",
